Add files via upload
This is the initial program where 5 speed level is utilized. Main problem is it will start on level 3 (around 153). Thus one of the solutions I can think of to start the motor at level 1 is to start the motor at level 3 or full speed then using loops it will go down to its coresponding number.